requirements.txt
certifi==2019.3.9
coverage==4.5.3
dj-database-url==0.5.0
Django==2.2
django-extensions==2.1.6
django-render-block==0.5
django-templated-email==2.3.0
djangorestframework==3.9.2
entrypoints==0.3
flake8==3.7.7
gunicorn==19.9.0
mccabe==0.6.1
psycopg2==2.8.2
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.1
sentry-sdk==0.8.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
uuid==1.30
whitenoise==4.1.2
